Moved relevant scenes into new file--cleaned up scene
Added small dots to the animation that indiciate the current inputs on the graph. IMPORTANTLY these two graphs do not properly correspond to each other, as Ijust wanted to get the animation up and running. Now the scenes that will be used are in MOSFETGraphingScenes.py, and the _learningScenes are just for reference for me, likely to be deleted soon. Next, the correspondence between the graphs needs to be established.