Input/output graph animations
Building from the last scene I created a new scene that after displaying the linear and saturation regions, as well as ON and OFF, shows how graphs are correlated. Then, these graphs are moved and two more graphs representing the rough input and output are plotted. Note that there is no specific circuit topology yet so it doesn't really correspond to anything specific yet (the K'(W/L) is also inaccurate for typical MOSFETs). Additionally, there is no labelling yet and the code could be cleaner.