Created small MOSFET circuit
The small MOSFET circuit was done somewhat clunkily I think circuitry objects could be made and used my building modular components for it instead of the way I have it set up. However, it does the job for now. Also added a scene that just stiches the two existing scenes (that are important) together. So, it shows the circuit and then shows the graphs. Still not much description for a viewer though.